+++ pcsx2/PAD/Gamepad.h
@@ -15,48 +15,21 @@
 
 #pragma once
 
-#include "Global.h"
-#include "Host.h"
-#include "common/mt_queue.h"
+#include "common/WindowInfo.h"
 #include "SaveState.h"
-
-enum PadOptions
-{
-	PADOPTION_FORCEFEEDBACK = 0x1,
-	PADOPTION_REVERSELX = 0x2,
-	PADOPTION_REVERSELY = 0x4,
-	PADOPTION_REVERSERX = 0x8,
-	PADOPTION_REVERSERY = 0x10,
-	PADOPTION_MOUSE_L = 0x20,
-	PADOPTION_MOUSE_R = 0x40,
-};
-
-struct WindowInfo;
-
-extern FILE* padLog;
-extern void initLogging();
-
-extern HostKeyEvent event;
-extern MtQueue<HostKeyEvent> g_ev_fifo;
-
-s32 _PADopen(const WindowInfo& wi);
-void _PADclose();
-void PADsetMode(int pad, int mode);
-
-void SysMessage(char* fmt, ...);
+#include "Host.h"
 
 s32 PADinit();
 void PADshutdown();
 s32 PADopen(const WindowInfo& wi);
-void PADsetLogDir(const char* dir);
+void PADupdate(int pad);
 void PADclose();
-s32 PADsetSlot(u8 port, u8 slot);
-s32 PADfreeze(FreezeAction mode, freezeData* data);
 u8 PADstartPoll(int pad);
 u8 PADpoll(u8 value);
 HostKeyEvent* PADkeyEvent();
-void PADupdate(int pad);
 void PADconfigure();
+s32 PADfreeze(FreezeAction mode, freezeData* data);
+s32 PADsetSlot(u8 port, u8 slot);
 
 #if defined(__unix__)
 void PADWriteEvent(HostKeyEvent& evt);
