+++ pcsx2/PAD/Linux/Global.h
@@ -30,6 +30,7 @@
 
 #include "common/pxStreams.h"
 #include "common/Console.h"
+#include "common/mt_queue.h"
 #include "DebugTools/Debug.h"
 
 #define PADdefs
@@ -93,6 +94,17 @@ static const std::array<gamePadValues, MAX_KEYS> all_keys =
 	PAD_R_RIGHT,  // Right joystick (Right) →
 	PAD_R_DOWN,   // Right joystick (Down) ↓
 	PAD_R_LEFT    // Right joystick (Left) ←
+};
+
+enum PadOptions
+{
+	PADOPTION_FORCEFEEDBACK = 0x1,
+	PADOPTION_REVERSELX = 0x2,
+	PADOPTION_REVERSELY = 0x4,
+	PADOPTION_REVERSERX = 0x8,
+	PADOPTION_REVERSERY = 0x10,
+	PADOPTION_MOUSE_L = 0x20,
+	PADOPTION_MOUSE_R = 0x40,
 };
 
 static bool IsAnalogKey(int index)
