+++ pcsx2/PAD/Linux/keyboard.cpp
@@ -20,7 +20,10 @@
 
 #include "Global.h"
 #include "keyboard.h"
-#include "PAD.h"
+#include "../Gamepad.h"
+
+extern HostKeyEvent event;
+extern MtQueue<HostKeyEvent> g_ev_fifo;
 
 /// g_key_status.press but with proper handling for analog buttons
 static void PressButton(u32 pad, u32 button)
