+++ pcsx2/SaveState.cpp
@@ -35,11 +35,7 @@
 #include "common/StringUtil.h"
 #include "SPU2/spu2.h"
 #include "USB/USB.h"
-#ifdef _WIN32
-#include "PAD/Windows/PAD.h"
-#else
-#include "PAD/Linux/PAD.h"
-#endif
+#include "PAD/Gamepad.h"
 
 #include <wx/wfstream.h>
