+++ pcsx2/Sio.cpp
@@ -19,11 +19,7 @@
 #include "Common.h"
 #include "Sio.h"
 #include "sio_internal.h"
-#ifdef _WIN32
-#include "PAD/Windows/PAD.h"
-#else
-#include "PAD/Linux/PAD.h"
-#endif
+#include "PAD/Gamepad.h"
 
 #ifndef DISABLE_RECORDING
 #	include "Recording/InputRecording.h"
