+++ pcsx2/gui/MainMenuClicks.cpp
@@ -23,11 +23,7 @@
 #include "System/SysThreads.h"
 #include "DEV9/DEV9.h"
 #include "USB/USB.h"
-#ifdef _WIN32
-#include "PAD/Windows/PAD.h"
-#else
-#include "PAD/Linux/PAD.h"
-#endif
+#include "PAD/Gamepad.h"
 
 #include "ConsoleLogger.h"
 #include "MainFrame.h"
