+++ pcsx2/gui/SysState.cpp
@@ -24,11 +24,7 @@
 #include "common/pxStreams.h"
 #include "SPU2/spu2.h"
 #include "USB/USB.h"
-#ifdef _WIN32
-#include "PAD/Windows/PAD.h"
-#else
-#include "PAD/Linux/PAD.h"
-#endif
+#include "PAD/Gamepad.h"
 
 #include "ConsoleLogger.h"
